modify purchase page
